refactor(core): split api-plugin endpoints into per-group route files
api-plugin.ts was 750+ lines of inline middleware handlers; scanning
"all endpoints" meant scrolling through every body. Move each prefix
group into its own routes/*.ts with a register*(server, ctx) function.
Each route file leads with a manifest comment of its endpoints, and
apiPlugin itself becomes a 30-line dispatcher whose body is literally
the list of route groups.

- routes/context.ts   ApiContext + readBody/json/resolveSlideEntryPath
- routes/watchers.ts  manifest + asset HMR pings
- routes/edit.ts      /__edit, /__edit/revert-asset, /__edit/batch
- routes/comments.ts  /__comments (list/add/delete)
- routes/slides.ts    /__slides/:id (PATCH/DELETE), reorder, pages/:i
- routes/assets.ts    /__assets/:scope CRUD + /usages
- routes/svgl.ts      /__svgl/search, /__svgl/svg
- routes/folders.ts   /__folders (list/create/assign/patch/delete)
